third_party/WebKit/LayoutTests/editing/inserting/multiple-lines-selected.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<div id="log"></div>
<script>
test(() => assert_selection(
    '<div contenteditable><div>f^oo</div><div>b|ar</div></div>',
    'insertParagraph',
    '<div contenteditable><div>f</div><div>|ar</div></div>'),
    'Insert a paragraph separator into a selection that spans multiple paragraphs');
</script>
